requirements.txt
asgiref==3.7.2
astroid==2.15.6
dill==0.3.7
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.10.0
pylint==2.17.5
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomlkit==0.12.1
typed-ast==1.5.5
validate-docbr==1.10.0
wrapt==1.15.0
